break simulator into parts, add readme
