+++ simulator/main.rs
@@ -1,14 +1,18 @@
+use generation::{Arbitrary, ArbitraryFrom};
 use limbo_core::{Connection, Database, File, OpenFlags, PlatformIO, Result, RowResult, IO};
+use model::table::{Column, Name, Table, Value};
+use model::query::{Insert, Predicate, Query, Select};
+use properties::{property_insert_select, property_select_all};
 use rand::prelude::*;
 use rand_chacha::ChaCha8Rng;
 use std::cell::RefCell;
-use std::fmt::Display;
-use std::ops::Deref;
 use std::rc::Rc;
 use std::sync::Arc;
 use tempfile::TempDir;
 
-use anarchist_readable_name_generator_lib::readable_name_custom;
+mod generation;
+mod properties;
+mod model;
 
 struct SimulatorEnv {
     opts: SimulatorOpts,
@@ -45,526 +49,7 @@ struct SimulatorOpts {
     page_size: usize,
 }
 
-trait Arbitrary {
-    fn arbitrary<R: Rng>(rng: &mut R) -> Self;
-}
 
-trait ArbitraryFrom<T> {
-    fn arbitrary_from<R: Rng>(rng: &mut R, t: &T) -> Self;
-}
-
-struct Table {
-    rows: Vec<Vec<Value>>,
-    name: String,
-    columns: Vec<Column>,
-}
-
-impl Arbitrary for Table {
-    fn arbitrary<R: Rng>(rng: &mut R) -> Self {
-        let name = Name::arbitrary(rng).0;
-        let columns = (1..rng.gen_range(1..128))
-            .map(|_| Column::arbitrary(rng))
-            .collect();
-        Table {
-            rows: Vec::new(),
-            name,
-            columns,
-        }
-    }
-}
-
-#[derive(Clone)]
-struct Column {
-    name: String,
-    column_type: ColumnType,
-    primary: bool,
-    unique: bool,
-}
-
-impl Arbitrary for Column {
-    fn arbitrary<R: Rng>(rng: &mut R) -> Self {
-        let name = Name::arbitrary(rng).0;
-        let column_type = ColumnType::arbitrary(rng);
-        Column {
-            name,
-            column_type,
-            primary: false,
-            unique: false,
-        }
-    }
-}
-
-#[derive(Clone)]
-enum ColumnType {
-    Integer,
-    Float,
-    Text,
-    Blob,
-}
-
-impl Arbitrary for ColumnType {
-    fn arbitrary<R: Rng>(rng: &mut R) -> Self {
-        match rng.gen_range(0..4) {
-            0 => ColumnType::Integer,
-            1 => ColumnType::Float,
-            2 => ColumnType::Text,
-            3 => ColumnType::Blob,
-            _ => unreachable!(),
-        }
-    }
-}
-
-#[derive(Clone, Debug, PartialEq)]
-enum Value {
-    Null,
-    Integer(i64),
-    Float(f64),
-    Text(String),
-    Blob(Vec<u8>),
-}
-
-impl ArbitraryFrom<Vec<&Value>> for Value {
-    fn arbitrary_from<R: Rng>(rng: &mut R, values: &Vec<&Value>) -> Self {
-        if values.is_empty() {
-            return Value::Null;
-        }
-
-        let index = rng.gen_range(0..values.len());
-        values[index].clone()
-    }
-}
-
-impl ArbitraryFrom<ColumnType> for Value {
-    fn arbitrary_from<R: Rng>(rng: &mut R, column_type: &ColumnType) -> Self {
-        match column_type {
-            ColumnType::Integer => Value::Integer(rng.gen_range(i64::MIN..i64::MAX)),
-            ColumnType::Float => Value::Float(rng.gen_range(-1e10..1e10)),
-            ColumnType::Text => Value::Text(gen_random_text(rng)),
-            ColumnType::Blob => Value::Blob(gen_random_text(rng).as_bytes().to_vec()),
-        }
-    }
-}
-
-struct LTValue(Value);
-
-impl ArbitraryFrom<Vec<&Value>> for LTValue {
-    fn arbitrary_from<R: Rng>(rng: &mut R, values: &Vec<&Value>) -> Self {
-        if values.is_empty() {
-            return LTValue(Value::Null);
-        }
-
-        let index = rng.gen_range(0..values.len());
-        LTValue::arbitrary_from(rng, values[index])
-    }
-}
-
-impl ArbitraryFrom<Value> for LTValue {
-    fn arbitrary_from<R: Rng>(rng: &mut R, value: &Value) -> Self {
-        match value {
-            Value::Integer(i) => LTValue(Value::Integer(rng.gen_range(i64::MIN..*i - 1))),
-            Value::Float(f) => LTValue(Value::Float(rng.gen_range(-1e10..*f - 1.0))),
-            Value::Text(t) => {
-                // Either shorten the string, or make at least one character smaller and mutate the rest
-                let mut t = t.clone();
-                if rng.gen_bool(0.01) {
-                    t.pop();
-                    LTValue(Value::Text(t))
-                } else {
-                    let mut t = t.chars().map(|c| c as u32).collect::<Vec<_>>();
-                    let index = rng.gen_range(0..t.len());
-                    t[index] -= 1;
-                    // Mutate the rest of the string
-                    for i in (index + 1)..t.len() {
-                        t[i] = rng.gen_range('a' as u32..='z' as u32);
-                    }
-                    let t = t
-                        .into_iter()
-                        .map(|c| char::from_u32(c).unwrap_or('z'))
-                        .collect::<String>();
-                    LTValue(Value::Text(t))
-                }
-            }
-            Value::Blob(b) => {
-                // Either shorten the blob, or make at least one byte smaller and mutate the rest
-                let mut b = b.clone();
-                if rng.gen_bool(0.01) {
-                    b.pop();
-                    LTValue(Value::Blob(b))
-                } else {
-                    let index = rng.gen_range(0..b.len());
-                    b[index] -= 1;
-                    // Mutate the rest of the blob
-                    for i in (index + 1)..b.len() {
-                        b[i] = rng.gen_range(0..=255);
-                    }
-                    LTValue(Value::Blob(b))
-                }
-            }
-            _ => unreachable!(),
-        }
-    }
-}
-
-struct GTValue(Value);
-
-impl ArbitraryFrom<Vec<&Value>> for GTValue {
-    fn arbitrary_from<R: Rng>(rng: &mut R, values: &Vec<&Value>) -> Self {
-        if values.is_empty() {
-            return GTValue(Value::Null);
-        }
-
-        let index = rng.gen_range(0..values.len());
-        GTValue::arbitrary_from(rng, values[index])
-    }
-}
-
-impl ArbitraryFrom<Value> for GTValue {
-    fn arbitrary_from<R: Rng>(rng: &mut R, value: &Value) -> Self {
-        match value {
-            Value::Integer(i) => GTValue(Value::Integer(rng.gen_range(*i..i64::MAX))),
-            Value::Float(f) => GTValue(Value::Float(rng.gen_range(*f..1e10))),
-            Value::Text(t) => {
-                // Either lengthen the string, or make at least one character smaller and mutate the rest
-                let mut t = t.clone();
-                if rng.gen_bool(0.01) {
-                    t.push(rng.gen_range(0..=255) as u8 as char);
-                    GTValue(Value::Text(t))
-                } else {
-                    let mut t = t.chars().map(|c| c as u32).collect::<Vec<_>>();
-                    let index = rng.gen_range(0..t.len());
-                    t[index] += 1;
-                    // Mutate the rest of the string
-                    for i in (index + 1)..t.len() {
-                        t[i] = rng.gen_range('a' as u32..='z' as u32);
-                    }
-                    let t = t
-                        .into_iter()
-                        .map(|c| char::from_u32(c).unwrap_or('a'))
-                        .collect::<String>();
-                    GTValue(Value::Text(t))
-                }
-            }
-            Value::Blob(b) => {
-                // Either lengthen the blob, or make at least one byte smaller and mutate the rest
-                let mut b = b.clone();
-                if rng.gen_bool(0.01) {
-                    b.push(rng.gen_range(0..=255));
-                    GTValue(Value::Blob(b))
-                } else {
-                    let index = rng.gen_range(0..b.len());
-                    b[index] += 1;
-                    // Mutate the rest of the blob
-                    for i in (index + 1)..b.len() {
-                        b[i] = rng.gen_range(0..=255);
-                    }
-                    GTValue(Value::Blob(b))
-                }
-            }
-            _ => unreachable!(),
-        }
-    }
-}
-
-enum Predicate {
-    And(Vec<Predicate>), // p1 AND p2 AND p3... AND pn
-    Or(Vec<Predicate>),  // p1 OR p2 OR p3... OR pn
-    Eq(String, Value),   // column = Value
-    Gt(String, Value),   // column > Value
-    Lt(String, Value),   // column < Value
-}
-
-// This type represents the potential queries on the database.
-enum Query {
-    Create(Create),
-    Select(Select),
-    Insert(Insert),
-    Delete(Delete),
-}
-
-struct Create {
-    table: Table,
-}
-
-impl Arbitrary for Create {
-    fn arbitrary<R: Rng>(rng: &mut R) -> Self {
-        Create {
-            table: Table::arbitrary(rng),
-        }
-    }
-}
-
-struct Select {
-    table: String,
-    predicate: Predicate,
-}
-
-impl ArbitraryFrom<Vec<Table>> for Select {
-    fn arbitrary_from<R: Rng>(rng: &mut R, tables: &Vec<Table>) -> Self {
-        let table = rng.gen_range(0..tables.len());
-        Select {
-            table: tables[table].name.clone(),
-            predicate: Predicate::arbitrary_from(rng, &tables[table]),
-        }
-    }
-}
-
-impl ArbitraryFrom<Vec<&Table>> for Select {
-    fn arbitrary_from<R: Rng>(rng: &mut R, tables: &Vec<&Table>) -> Self {
-        let table = rng.gen_range(0..tables.len());
-        Select {
-            table: tables[table].name.clone(),
-            predicate: Predicate::arbitrary_from(rng, tables[table]),
-        }
-    }
-}
-
-struct Insert {
-    table: String,
-    values: Vec<Value>,
-}
-
-impl ArbitraryFrom<Table> for Insert {
-    fn arbitrary_from<R: Rng>(rng: &mut R, table: &Table) -> Self {
-        let values = table
-            .columns
-            .iter()
-            .map(|c| Value::arbitrary_from(rng, &c.column_type))
-            .collect();
-        Insert {
-            table: table.name.clone(),
-            values,
-        }
-    }
-}
-
-struct Delete {
-    table: String,
-    predicate: Predicate,
-}
-
-impl ArbitraryFrom<Table> for Delete {
-    fn arbitrary_from<R: Rng>(rng: &mut R, table: &Table) -> Self {
-        Delete {
-            table: table.name.clone(),
-            predicate: Predicate::arbitrary_from(rng, table),
-        }
-    }
-}
-
-impl ArbitraryFrom<Table> for Query {
-    fn arbitrary_from<R: Rng>(rng: &mut R, table: &Table) -> Self {
-        match rng.gen_range(0..=200) {
-            0 => Query::Create(Create::arbitrary(rng)),
-            1..=100 => Query::Select(Select::arbitrary_from(rng, &vec![table])),
-            101..=200 => Query::Insert(Insert::arbitrary_from(rng, table)),
-            // todo: This branch is currently never taken, as DELETE is not yet implemented.
-            //       Change this when DELETE is implemented.
-            201..=300 => Query::Delete(Delete::arbitrary_from(rng, table)),
-            _ => unreachable!(),
-        }
-    }
-}
-
-struct CompoundPredicate(Predicate);
-struct SimplePredicate(Predicate);
-
-impl ArbitraryFrom<(&Table, bool)> for SimplePredicate {
-    fn arbitrary_from<R: Rng>(rng: &mut R, (table, predicate_value): &(&Table, bool)) -> Self {
-        // Pick a random column
-        let column_index = rng.gen_range(0..table.columns.len());
-        let column = &table.columns[column_index];
-        let column_values = table
-            .rows
-            .iter()
-            .map(|r| &r[column_index])
-            .collect::<Vec<_>>();
-        // Pick an operator
-        let operator = match rng.gen_range(0..3) {
-            0 => {
-                if *predicate_value {
-                    Predicate::Eq(
-                        column.name.clone(),
-                        Value::arbitrary_from(rng, &column_values),
-                    )
-                } else {
-                    Predicate::Eq(
-                        column.name.clone(),
-                        Value::arbitrary_from(rng, &column.column_type),
-                    )
-                }
-            }
-            1 => Predicate::Gt(
-                column.name.clone(),
-                match predicate_value {
-                    true => GTValue::arbitrary_from(rng, &column_values).0,
-                    false => LTValue::arbitrary_from(rng, &column_values).0,
-                },
-            ),
-            2 => Predicate::Lt(
-                column.name.clone(),
-                match predicate_value {
-                    true => LTValue::arbitrary_from(rng, &column_values).0,
-                    false => GTValue::arbitrary_from(rng, &column_values).0,
-                },
-            ),
-            _ => unreachable!(),
-        };
-
-        SimplePredicate(operator)
-    }
-}
-
-impl ArbitraryFrom<(&Table, bool)> for CompoundPredicate {
-    fn arbitrary_from<R: Rng>(rng: &mut R, (table, predicate_value): &(&Table, bool)) -> Self {
-        // Decide if you want to create an AND or an OR
-        CompoundPredicate(if rng.gen_bool(0.7) {
-            // An AND for true requires each of its children to be true
-            // An AND for false requires at least one of its children to be false
-            if *predicate_value {
-                Predicate::And(
-                    (0..rng.gen_range(1..=3))
-                        .map(|_| SimplePredicate::arbitrary_from(rng, &(*table, true)).0)
-                        .collect(),
-                )
-            } else {
-                // Create a vector of random booleans
-                let mut booleans = (0..rng.gen_range(1..=3))
-                    .map(|_| rng.gen_bool(0.5))
-                    .collect::<Vec<_>>();
-
-                let len = booleans.len();
-
-                // Make sure at least one of them is false
-                if booleans.iter().all(|b| *b) {
-                    booleans[rng.gen_range(0..len)] = false;
-                }
-
-                Predicate::And(
-                    booleans
-                        .iter()
-                        .map(|b| SimplePredicate::arbitrary_from(rng, &(*table, *b)).0)
-                        .collect(),
-                )
-            }
-        } else {
-            // An OR for true requires at least one of its children to be true
-            // An OR for false requires each of its children to be false
-            if *predicate_value {
-                // Create a vector of random booleans
-                let mut booleans = (0..rng.gen_range(1..=3))
-                    .map(|_| rng.gen_bool(0.5))
-                    .collect::<Vec<_>>();
-                let len = booleans.len();
-                // Make sure at least one of them is true
-                if booleans.iter().all(|b| !*b) {
-                    booleans[rng.gen_range(0..len)] = true;
-                }
-
-                Predicate::Or(
-                    booleans
-                        .iter()
-                        .map(|b| SimplePredicate::arbitrary_from(rng, &(*table, *b)).0)
-                        .collect(),
-                )
-            } else {
-                Predicate::Or(
-                    (0..rng.gen_range(1..=3))
-                        .map(|_| SimplePredicate::arbitrary_from(rng, &(*table, false)).0)
-                        .collect(),
-                )
-            }
-        })
-    }
-}
-
-impl ArbitraryFrom<Table> for Predicate {
-    fn arbitrary_from<R: Rng>(rng: &mut R, table: &Table) -> Self {
-        let predicate_value = rng.gen_bool(0.5);
-        CompoundPredicate::arbitrary_from(rng, &(table, predicate_value)).0
-    }
-}
-
-impl ArbitraryFrom<(&str, &Value)> for Predicate {
-    fn arbitrary_from<R: Rng>(rng: &mut R, (column_name, value): &(&str, &Value)) -> Self {
-        match rng.gen_range(0..3) {
-            0 => Predicate::Eq(column_name.to_string(), (*value).clone()),
-            1 => Predicate::Gt(
-                column_name.to_string(),
-                LTValue::arbitrary_from(rng, *value).0,
-            ),
-            2 => Predicate::Lt(
-                column_name.to_string(),
-                LTValue::arbitrary_from(rng, *value).0,
-            ),
-            _ => unreachable!(),
-        }
-    }
-}
-
-impl Display for Predicate {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Predicate::And(predicates) => {
-                if predicates.is_empty() {
-                    // todo: Make this TRUE when the bug is fixed
-                    write!(f, "TRUE")
-                } else {
-                    write!(f, "(")?;
-                    for (i, p) in predicates.iter().enumerate() {
-                        if i != 0 {
-                            write!(f, " AND ")?;
-                        }
-                        write!(f, "{}", p)?;
-                    }
-                    write!(f, ")")
-                }
-            }
-            Predicate::Or(predicates) => {
-                if predicates.is_empty() {
-                    write!(f, "FALSE")
-                } else {
-                    write!(f, "(")?;
-                    for (i, p) in predicates.iter().enumerate() {
-                        if i != 0 {
-                            write!(f, " OR ")?;
-                        }
-                        write!(f, "{}", p)?;
-                    }
-                    write!(f, ")")
-                }
-            }
-            Predicate::Eq(name, value) => write!(f, "{} = {}", name, value),
-            Predicate::Gt(name, value) => write!(f, "{} > {}", name, value),
-            Predicate::Lt(name, value) => write!(f, "{} < {}", name, value),
-        }
-    }
-}
-
-impl Display for Query {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Query::Create(Create { table }) => write!(f, "{}", table.to_create_str()),
-            Query::Select(Select {
-                table,
-                predicate: guard,
-            }) => write!(f, "SELECT * FROM {} WHERE {}", table, guard),
-            Query::Insert(Insert { table, values }) => {
-                write!(f, "INSERT INTO {} VALUES (", table)?;
-                for (i, v) in values.iter().enumerate() {
-                    if i != 0 {
-                        write!(f, ", ")?;
-                    }
-                    write!(f, "{}", v)?;
-                }
-                write!(f, ")")
-            }
-            Query::Delete(Delete {
-                table,
-                predicate: guard,
-            }) => write!(f, "DELETE FROM {} WHERE {}", table, guard),
-        }
-    }
-}
 
 #[allow(clippy::arc_with_non_send_sync)]
 fn main() {
@@ -652,68 +137,6 @@ fn main() {
     env.io.print_stats();
 }
 
-fn property_insert_select(env: &mut SimulatorEnv, conn: &mut Rc<Connection>) {
-    // Get a random table
-    let table = env.rng.gen_range(0..env.tables.len());
-
-    // Pick a random column
-    let column_index = env.rng.gen_range(0..env.tables[table].columns.len());
-    let column = &env.tables[table].columns[column_index].clone();
-
-    let mut rng = env.rng.clone();
-
-    // Generate a random value of the column type
-    let value = Value::arbitrary_from(&mut rng, &column.column_type);
-
-    // Create a whole new row
-    let mut row = Vec::new();
-    for (i, column) in env.tables[table].columns.iter().enumerate() {
-        if i == column_index {
-            row.push(value.clone());
-        } else {
-            let value = Value::arbitrary_from(&mut rng, &column.column_type);
-            row.push(value);
-        }
-    }
-
-    // Insert the row
-    let query = Query::Insert(Insert {
-        table: env.tables[table].name.clone(),
-        values: row.clone(),
-    });
-    let _ = get_all_rows(env, conn, query.to_string().as_str()).unwrap();
-    // Shadow operation on the table
-    env.tables[table].rows.push(row.clone());
-
-    // Create a query that selects the row
-    let query = Query::Select(Select {
-        table: env.tables[table].name.clone(),
-        predicate: Predicate::Eq(column.name.clone(), value),
-    });
-
-    // Get all rows
-    let rows = get_all_rows(env, conn, query.to_string().as_str()).unwrap();
-
-    // Check that the row is there
-    assert!(rows.iter().any(|r| r == &row));
-}
-
-fn property_select_all(env: &mut SimulatorEnv, conn: &mut Rc<Connection>) {
-    // Get a random table
-    let table = env.rng.gen_range(0..env.tables.len());
-
-    // Create a query that selects all rows
-    let query = Query::Select(Select {
-        table: env.tables[table].name.clone(),
-        predicate: Predicate::And(Vec::new()),
-    });
-
-    // Get all rows
-    let rows = get_all_rows(env, conn, query.to_string().as_str()).unwrap();
-
-    // Make sure the rows are the same
-    compare_equal_rows(&rows, &env.tables[table].rows);
-}
 
 fn process_connection(env: &mut SimulatorEnv, conn: &mut Rc<Connection>) -> Result<()> {
     if env.tables.is_empty() {
@@ -827,38 +250,7 @@ fn maybe_add_table(env: &mut SimulatorEnv, conn: &mut Rc<Connection>) -> Result<
     Ok(())
 }
 
-struct Name(String);
 
-impl Arbitrary for Name {
-    fn arbitrary<R: Rng>(rng: &mut R) -> Self {
-        let name = readable_name_custom("_", rng);
-        Name(name.replace("-", "_"))
-    }
-}
-
-impl Deref for Name {
-    type Target = str;
-
-    fn deref(&self) -> &Self::Target {
-        &self.0
-    }
-}
-
-fn gen_random_text<T: Rng>(rng: &mut T) -> String {
-    let big_text = rng.gen_ratio(1, 1000);
-    if big_text {
-        let max_size: u64 = 2 * 1024 * 1024 * 1024;
-        let size = rng.gen_range(1024..max_size);
-        let mut name = String::new();
-        for i in 0..size {
-            name.push(((i % 26) as u8 + b'A') as char);
-        }
-        name
-    } else {
-        let name = readable_name_custom("_", rng);
-        name.replace("-", "_")
-    }
-}
 
 fn get_all_rows(
     env: &mut SimulatorEnv,
@@ -1086,48 +478,3 @@ impl Drop for SimulatorFile {
     }
 }
 
-impl ColumnType {
-    pub fn as_str(&self) -> &str {
-        match self {
-            ColumnType::Integer => "INTEGER",
-            ColumnType::Float => "FLOAT",
-            ColumnType::Text => "TEXT",
-            ColumnType::Blob => "BLOB",
-        }
-    }
-}
-
-impl Table {
-    pub fn to_create_str(&self) -> String {
-        let mut out = String::new();
-
-        out.push_str(format!("CREATE TABLE {} (", self.name).as_str());
-
-        assert!(!self.columns.is_empty());
-        for column in &self.columns {
-            out.push_str(format!("{} {},", column.name, column.column_type.as_str()).as_str());
-        }
-        // remove last comma
-        out.pop();
-
-        out.push_str(");");
-        out
-    }
-}
-
-impl Display for Value {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        match self {
-            Value::Null => write!(f, "NULL"),
-            Value::Integer(i) => write!(f, "{}", i),
-            Value::Float(fl) => write!(f, "{}", fl),
-            Value::Text(t) => write!(f, "'{}'", t),
-            Value::Blob(b) => write!(f, "{}", to_sqlite_blob(b)),
-        }
-    }
-}
-
-fn to_sqlite_blob(bytes: &[u8]) -> String {
-    let hex: String = bytes.iter().map(|b| format!("{:02X}", b)).collect();
-    format!("X'{}'", hex)
-}
